+++ docs/snippets/resilience/circuitbreaker_redis.py
@@ -1,8 +1,8 @@
 from grelmicro import Grelmicro
 from grelmicro.providers.redis import RedisProvider
-from grelmicro.resilience import CircuitBreaker, CircuitBreakerRegistry
+from grelmicro.resilience import CircuitBreaker
 
 redis = RedisProvider("redis://localhost:6379/0")
-micro = Grelmicro(uses=[CircuitBreakerRegistry(redis)])
+micro = Grelmicro(uses=[redis])
 
 payments = CircuitBreaker("payments")